+++ src/CoreLib/Game/Services/MoveService.cs
@@ -212,7 +212,8 @@ internal class MoveService : MessageService, IWithTimers {
     private void SendZoneInteractionFishRequest() {
         var msg = new ZONE_102_PROTOCOL.MSG_PLAYERMOVE() {
             PlayerObject = _activeCoreObject,
-            PlayerActor = SessionActor.ActorRef
+            PlayerActor = SessionActor.ActorRef,
+            PlayerWizard = _wizard,
         };
 
         TellOtherServices(msg);
