+++ src/CoreLib/Game/Zone/Components/InteractTrainerComponent.cs
@@ -17,7 +17,7 @@ using static Imlight.Common.Caches.TypeCache;
 
 namespace Imlight.CoreLib.Game.Zone.Components;
 
-internal sealed class TrainerComponent(ZoneEntity entity) : ZoneEntityComponent(entity), IServiceComponent, IComponentFactory {
+internal sealed class InteractTrainerComponent(ZoneEntity entity) : ZoneEntityComponent(entity), IServiceComponent, IComponentFactory {
 
     public string ServiceName     => "WizTrainingService";
     public string NpcIcon         => null;
