+++ src/CoreLib/Game/Zone/Components/InteractVendorComponent.cs
@@ -4,38 +4,60 @@
  */
 
 using Akka.Actor;
+using Imlight.Common;
 using Imlight.Common.Caches;
 using Imlight.Common.Cryptography;
+using Imlight.Common.ObjectProperty;
+using Imlight.Common.ObjectProperty.PropertyReflection;
+using Imlight.CoreLib.Game.World;
 using Imlight.CoreLib.Game.Zone.Core;
 using Imlight.CoreLib.Shared.Packets;
+using Imlight.CoreLib.WizardData.Collections;
 using Imlight.CoreLib.WizardData.Models.Player;
 using System.Collections.Generic;
+using System.Linq;
 using static Imlight.Common.Caches.TypeCache;
 
 namespace Imlight.CoreLib.Game.Zone.Components;
 
-internal sealed class DyeShopComponent(ZoneEntity entity) : ZoneEntityComponent(entity), IServiceComponent, IComponentFactory {
+internal sealed class InteractVendorComponent(ZoneEntity entity) : ZoneEntityComponent(entity), IServiceComponent, IComponentFactory {
 
-    // Dye shop NPCs always have "dye" in their object name. It's what we can use to deduce if the entity is a dye shop NPC.
-    private const string DYE_SHOP_NPC_CONTAINS = "Dye";
-    private const string DYE_SHOP_TITLE = "WC-NPCs_00000718";
-
-    public string ServiceName     => "DyeShopService";
+    public string ServiceName     => "WizShoppingService";
     public string NpcIcon         => null;
     public string NpcNameKey      => null;
     public string NpcTextKey      => null;
     public string WizBang         => "Shopping";
     public string StateName       => "Shop";
     public string InteractWizBang => "Registrar";
-    public string DisplayKey      => "GUI_DyeShop";
+    public string DisplayKey      => "GUI_ShopOptionEquipment";
+
+    private static readonly ObjectSerializer s_offeringsSerializer = new ObjectSerializer()
+            .OnBehaviors(SerializerOptions.Behaviors.None)
+            .OnPropertyMask((SerializerOptions.PropertyFlags) 4);
+    private List<GID> _inventory;
 
     public static bool ShouldAttachToEntity(CoreTemplate template) 
-        => template is GameObjectTemplate gameObjectTemplate
-        && gameObjectTemplate.m_objectName.ToString().Contains(DYE_SHOP_NPC_CONTAINS);
+        // Attach if the template is an NPC and has an inventory in Dragon database,
+        // or if the template is a vendor as per game client data.
+        => template is GameObjectTemplate goTemplate 
+        && goTemplate.m_behaviors.Any(x => x is NPCBehaviorTemplate) 
+        && (NpcInventoryCollection.TryGetNpcInventory(goTemplate.m_templateID, out _)
+        || WorldVendorLocations.IsVendor(goTemplate.m_templateID));
+
+    public override void OnStart() {
+        if (!NpcInventoryCollection.TryGetNpcInventory(Entity.ActiveGameObject.m_templateID, out var inventory)) {
+            Logger.Error("Failed to get vendor inventory for NPC {0}", 
+                Logger.Args(Entity.ActiveGameObject.m_templateID));
+
+            return;
+        }
+
+        _inventory = inventory.Inventory;
+    }
 
     public IEnumerable<ServiceOptionBase> GetServiceOptions(Wizard _) 
         => [
-            new DyeShopOption {
+            new EquipmentShopOption {
                 m_displayKey = DisplayKey,
                 m_iconKey = NpcIcon,
                 m_serviceName = ServiceName,
@@ -43,18 +65,31 @@ internal sealed class DyeShopComponent(ZoneEntity entity) : ZoneEntityComponent(
         ];
 
     public void OnServiceInteraction(IActorRef playerActor, Wizard playerCharacter, CoreObject playerObject, uint serviceOptionIndex) {
-        SendPlayerDyeShopOpen(playerActor, Entity.ActiveGameObject.m_globalID);
+        SendShopOfferings(playerActor);
         SendPlayerIntoWizbang(playerObject.m_globalID);
         SendPlayerIntoState(playerObject.m_globalID);
     }
 
-    private void SendPlayerDyeShopOpen(IActorRef playerActor, ulong objId) {
-        var dyeShopOpen = new WIZARD_12_PROTOCOL.MSG_DYESHOPOPEN() {
-            GlobalID = objId,
-            Title = DYE_SHOP_TITLE
-        };
+    public bool HasItem(GID itemGID) 
+        => _inventory.Any(x => x.MParts.Id == itemGID.MParts.Id);
 
-        playerActor.Tell(dyeShopOpen);
+    private void SendShopOfferings(IActorRef playerActor) {
+        var shopOffering = new WizShopOffering() {
+            m_sellModifier = 0.05f,
+            m_shopTitle = "KrocNPC_00000013",
+            m_shopList = _inventory,
+
+            // Changes the type of currency that is used
+            // 0 - Gold
+            // 1 - PvP tickets
+            m_shopType = 0,
+        };
+        var data = s_offeringsSerializer.Serialize(shopOffering);
+        var shopListMsg = new WIZARD_12_PROTOCOL.MSG_SHOPLIST() {
+            GlobalID = Entity.ActiveGameObject.m_globalID,
+            Data = data,
+        };
+        playerActor.Tell(shopListMsg);
     }
 
     private void SendPlayerIntoWizbang(ulong playerObjID) {
@@ -84,5 +119,5 @@ internal sealed class DyeShopComponent(ZoneEntity entity) : ZoneEntityComponent(
 
         Entity.ZoneRef.Tell(broadcastMsg);
     }
-    
+
 }
