+++ src/CoreLib/Game/Zone/Components/RenderComponent.cs
@@ -90,7 +90,7 @@ internal sealed class RenderComponent(ZoneEntity entity) : ZoneEntityComponent(e
         DespawnObjectForPlayer(suspect);
     }
 
-    public override void OnPlayerMove(CoreObject playerObj, IActorRef playerActor) {
+    public override void OnPlayerMove(CoreObject playerObj, IActorRef playerActor, Wizard playerWizard) {
         if (!_doesDistanceCheck) {
             return;
         }
